+++ Assets/GameController.cs
@@ -8,7 +8,7 @@ public class GameController : MonoBehaviour
     [SerializeField] public float r = 5;
     [Range(0f, 1f)]
     [SerializeField] public float colorNum = 0.2f;
-    [Range(1, 35)]
+    [Range(1, 36)]
     [SerializeField] public int numOfSquares = 1;
     //[Range(1, 4)]
     //[SerializeField] private int span = 1;
@@ -18,13 +18,16 @@ public class GameController : MonoBehaviour
 
     [SerializeField] private OrbitMotion orbitMotion;
     [SerializeField] private ShapeGenerator shapeGenerator;
+    [SerializeField] private InputManager inputManager;
 
     private Color color = Color.red;
     private bool endOfGame = false;
 
 
+
     void Update()
     {
+
         if (Input.GetKey("space"))
         {
             spawn = false;
@@ -33,16 +36,32 @@ public class GameController : MonoBehaviour
         }
         if (!endOfGame)
         {
-            UpdateColor();
+            Vector4 data = inputManager.GetData();
             orbitMotion.orbitPath.r = r;
-            shapeGenerator.color = color;
-            shapeGenerator.respawnTime = respawnTime;
+            if (data[0] == 1f)
+            {
+                spawn = true;
+            }
+            else
+            {
+                spawn = true;
+            }
             shapeGenerator.spawn = spawn;
+
+            colorNum = data[1] / 360;
+            UpdateColor();
+            shapeGenerator.color = color;
+
+            numOfSquares = (int)(data[2] / 36) + 1;
+            shapeGenerator.respawnTime = (data[3] / 450) + 0.2f;
+
+
         }
         else
         {
             shapeGenerator.spawn = false;
         }
+        
     }
 
     void UpdateColor()
