+++ Assets/InputManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9d57820101312fe4581b49ceff3473f7
+guid: a5784baa21f400a4a8f3ffe60587cdba
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
