--- Assets/Interface.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a5784baa21f400a4a8f3ffe60587cdba
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
